JavaDoc for Servce and Controller classes

+++ src/main/java/co/uk/f3/payment/service/impl/PaymentServiceImpl.java
@@ -14,6 +14,12 @@ import co.uk.f3.payment.model.domain.Payment;
 import co.uk.f3.payment.repository.IPaymentRepository;
 import co.uk.f3.payment.service.IPaymentService;
 
+/**
+ * Class (Service) to allow execution of business logic operations.
+ * 
+ * @author Nissi Tafie
+ *
+ */
 @Service
 public class PaymentServiceImpl implements IPaymentService {
 
@@ -27,30 +33,70 @@ public class PaymentServiceImpl implements IPaymentService {
 
 	}
 
+	/**
+	 * Persists new Payment
+	 * 
+	 * @param to
+	 *            be persisted payment
+	 * @return created Payment
+	 */
 	@Override
 	public Optional<Payment> savePayment(Payment payment) {
 		return Optional.of(paymentRepository.insert(payment));
 	}
 
+	/**
+	 * Updates existing Payment
+	 * 
+	 * @param to
+	 *            be updated payment
+	 * @return updated Payment
+	 */
 	@Override
 	public Optional<Payment> updatePayment(Payment payment) {
 		return Optional.of(paymentRepository.save(payment));
 	}
 
+	/**
+	 * Fetches existing Payment by Id
+	 * 
+	 * @param id
+	 *            of existing payment
+	 * @return Payment
+	 */
 	@Override
 	public Optional<Payment> fetchPaymentById(String id) {
 		LOGGER.info(id);
-		return Optional.of(paymentRepository.findOne(id));
+		if (paymentRepository.exists(id)) {
+			return Optional.of(paymentRepository.findOne(id));
+		}
+		return Optional.empty();
 	}
 
+	/**
+	 * Fetches all existing payments.
+	 * 
+	 * @return list of existing Payments
+	 */
 	@Override
 	public Set<Payment> fetchPayments() {
 		return paymentRepository.findAll().stream().collect(Collectors.toSet());
 	}
 
+	/**
+	 * Deletes existing Payment by Id
+	 * 
+	 * @param id
+	 *            of existing payment
+	 * @return result of operation
+	 */
 	@Override
-	public void deletePaymentById(String id) {
-		paymentRepository.delete(id);
+	public boolean deletePaymentById(String id) {
+		if (paymentRepository.exists(id)) {
+			paymentRepository.delete(id);
+			return true;
+		}
+		return false;
 
 	}
